refactor(imports): Update imports in action folder
